pyproject.toml
[project]
name = "zerochat"
version = "0.2.0"
description = "A dead-simple command-line chat server and client using ZeroMQ"
readme = "README.md"
license = { file = "LICENSE.txt" }
requires-python = ">=3.10"
dependencies = [
    "aioconsole>=0.8.2",
    "pyzmq>=26.2",
    "rich>=13.9",
]

[project.scripts]
zerochat-server = "zerochat.server:main"
zerochat-client = "zerochat.client:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "black>=24.10",
    "flake8>=7.1",
    "isort>=5.13",
    "mypy>=1.13",
    "pre-commit>=4.0",
    "pytest>=8.3",
    "pytest-cov>=6.0",
]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
